refactor: select refactor - Moved functions and methods to utils
